% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/program.R
\name{assign_ous}
\alias{assign_ous}
\title{Assign OUS to a program}
\usage{
assign_ous(df = NULL, program_id = "program_id", live = F)
}
\arguments{
\item{df}{A data.frame with organisationUnit ids}

\item{program_id}{The id of the Program to update.}

\item{live}{A logical input (TRUE/FALSE) specifying the change
or where to apply to the changes. The default option is FALSE;
the changes will be made at the test server. You can specify this option
 to TRUE to apply the changes in production.}
}
\value{
A response
}
\description{
Update the organisation units of a program.
}
\details{
\code{assign_ous} updates the existing program ous with new organisation units (ous).
}
\examples{
\dontrun{
assign_ous(df, program_id = "BCnz7XE4olS")
assign_ous(df, program_id = "BCnz7XE4olS", live = TRUE)
}
}
