% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/program.R
\name{assign_ous_to_programs}
\alias{assign_ous_to_programs}
\title{Assign OUS to a program}
\usage{
assign_ous_to_programs(x, program_id = "program_id", live = F)
}
\arguments{
\item{x}{A data.frame, organisationUnits}

\item{program_id}{Program to assign}
}
\description{
Assign OUS to a program
}
