% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chop.R
\name{chop_sp}
\alias{chop_sp}
\title{Chop Attendance sheets}
\usage{
chop_sp(dt = NULL)
}
\arguments{
\item{dt}{A list of data.tables or data.frames}
}
\value{
transformed list
}
\description{
\code{chop_att} transforms attendance sheets in preparation for the next stage.
}
