% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/option_set.R
\name{de_option_set}
\alias{de_option_set}
\title{Data element option set}
\usage{
de_option_set(uid = NULL)
}
\arguments{
\item{uid}{A character string, the data element uid.}
}
\value{
an S3 type object with an optsion set id, options, path and the response.
}
\description{
Retrieve an option set of a data element.
}
\details{
\code{de_option_set} retrieves an option set of a data element. It checks whether the data element has
    an option set and extracts its details including the options. \code{de_option_set} returns the option set id, options,
    path and the response.
}
\examples{
\dontrun{
de_option_set(uid = "xywg23thsiw")
}
}
