% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{de_options}
\alias{de_options}
\title{GET options in a optionSet}
\usage{
de_options(uid = NULL)
}
\arguments{
\item{uid}{An optionSet uid}
}
\value{
an S3 object with the content, path and the response.
}
\description{
GET options in a optionSet
}
