% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{de_optionset}
\alias{de_optionset}
\title{GET optionSet details of a data element}
\usage{
de_optionset(uid = NULL)
}
\arguments{
\item{uid}{A dataElement uid}
}
\value{
an S3 object with the optsionSet id, options, path and the response.
}
\description{
GET optionSet details of a data element
}
