% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzzzz.R
\name{dt_names}
\alias{dt_names}
\title{Get data frame names}
\usage{
dt_names(dt)
}
\arguments{
\item{dt}{A list of data.table or data.frames}
}
\value{
A vector of characters
}
\description{
Get data frame names
}
