% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{event_dates}
\alias{event_dates}
\title{Event dates}
\usage{
event_dates(x)
}
\arguments{
\item{x}{An vector or character string}
}
\value{
DHIS2 eventDate
}
\description{
Convert elements dates to standard DHIS2 eventDates
}
