% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/payload.R
\name{format_payload}
\alias{format_payload}
\title{Format payload}
\usage{
format_payload(dt)
}
\arguments{
\item{dt}{A data.table or data.frame}
}
\value{
Event payload
}
\description{
generate an event payload from a dataframe object.
}
