% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{invalid_ous}
\alias{invalid_ous}
\title{Validate mapped ous}
\usage{
invalid_ous(dt)
}
\arguments{
\item{dt}{a list of data.tables or data.frames}
}
\value{
Invalid list of dfs
}
\description{
Validate mapped ous
}
