% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_files.R
\name{load_files_att2}
\alias{load_files_att2}
\title{Load A360 NG legacy files into R}
\usage{
load_files_att2(path = NULL)
}
\arguments{
\item{path}{System path to the directory with data}
}
\value{
a list of data frames
}
\description{
\code{load_files} loads A360 legacy files in the specified directory into R.
}
\details{
#'@seealso
* [load_files] Loads all the csv files in the specified directory.
* [load_files_att] Loads all the Attendance sheets in the specified directory.
* [load_sp] Loads all the service provision sheets in the specified directory.
}
\examples{
\dontrun{
load_files(file.path(Sys.getenv("HOME"),"data")
}
}
