% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{map_att}
\alias{map_att}
\title{Map Attendance sheets}
\usage{
map_att(dt = NULL)
}
\arguments{
\item{df}{A data.table or data.frame}
}
\value{
A mapped data.table or data.frame
}
\description{
\code{map_att} maps attendance sheets with the existing options codes in
NG RH A360 -Provider Client Records program in DHIS2.
}
