% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_ous.R
\name{map_clinics}
\alias{map_clinics}
\title{Map Clinics
map clinics in the files with exising one in DHIS2.}
\usage{
map_clinics(dt, ng_ous)
}
\arguments{
\item{dt}{A vector of character or string to modify}
}
\value{
uid of the mapped clinic
}
\description{
Map Clinics
map clinics in the files with exising one in DHIS2.
}
