% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{map_current_method}
\alias{map_current_method}
\title{Map Current Method}
\usage{
map_current_method(x = NULL)
}
\arguments{
\item{x}{A vector of character or string}
}
\value{
Existing option codes in DHIS2
}
\description{
\code{map_current_method} maps current methods with the options codes provided in
NG RH A360 - Provider Client Records program in DHIS2.
}
