% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{map_method_recieved}
\alias{map_method_recieved}
\title{Map Method Recieved}
\usage{
map_method_recieved(x = NULL)
}
\arguments{
\item{x}{A vector of character or string}
}
\value{
Existing option codes in DHIS2
}
\description{
\code{map_method_recieved} maps method recieved option with the option codes provided in
NG RH A360 - Provider Client Records program in DHIS2.
}
