% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/payload.R
\name{map_payload}
\alias{map_payload}
\title{Map payload}
\usage{
map_payload(x)
}
\arguments{
\item{x}{A list of data.frame or data.tables}
}
\value{
A mapped lsit of data.frames
}
\description{
Apply payload to a list of data.frames or data.tables
}
