% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{map_preg_results}
\alias{map_preg_results}
\title{Map pregenancy results}
\usage{
map_preg_results(x = NULL)
}
\arguments{
\item{x}{A vector of character or string}
}
\value{
Existing option codes in DHIS2
}
\description{
\code{map_preg_results} maps pregnancy results with the option codes provided in
NG RH A360 - Provider Client Records program in DHIS2.
}
