% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{map_program_activity}
\alias{map_program_activity}
\title{Map Program Activity}
\usage{
map_program_activity(x = NULL)
}
\arguments{
\item{x}{A vector of character or string}
}
\value{
Existing option codes in DHIS2
}
\description{
\code{map_program_entry_point} maps program activities with the option codes provided in
NG RH A360 - Provider Client Records program in DHIS2.
}
