% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{map_program_entry_point}
\alias{map_program_entry_point}
\title{Map Program Entry Point}
\usage{
map_program_entry_point(x = NULL)
}
\arguments{
\item{x}{A vector of character or string}
}
\value{
Existing option codes in DHIS2
}
\description{
\code{map_program_entry_point} maps program entry points with the option codes provided in
NG RH A360 - Provider Client Records program in DHIS2.
}
