% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{map_sp}
\alias{map_sp}
\title{Map service provision sheet}
\usage{
map_sp(df)
}
\arguments{
\item{df}{A data.table or data.frame}
}
\value{
A mapped data.table or data.frame
}
\description{
\code{map_sp} maps the service provision sheets with the existing options codes
provided in NG RH A360 - Provider Client Records program in DHIS2.
}
