% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{map_used_ec_last_sex}
\alias{map_used_ec_last_sex}
\title{Map used EC/Condom last sex}
\usage{
map_used_ec_last_sex(x = NULL)
}
\arguments{
\item{x}{A vector of character or string}
}
\value{
Existing option codes in DHIS2
}
\description{
\code{map_used_ec_last_sex} maps used EC/Condom last sex with the options codes provided in
NG RH A360 - Provider Client Records program in DHIS2.
}
