% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/program.R
\name{prog_des_9ja}
\alias{prog_des_9ja}
\title{GET NG RH A360 - A2. 9ja Girls & MMA Events program data elements}
\usage{
prog_des_9ja()
}
\value{
An S3 object
}
\description{
\code{prog_des_9ja} Retrieves all the data elements from NG RH A360 - A2. 9ja Girls & MMA Events program.
}
