% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/program.R
\name{prog_des_prov}
\alias{prog_des_prov}
\title{GET NG RH A360 - A3. Provider Client Record program data elements}
\usage{
prog_des_prov()
}
\value{
An S3 object
}
\description{
\code{prog_des_prov} Retrieves all the data elements from NG RH A360 - A3. Provider Client Record program
}
