% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{remove_empty_rows}
\alias{remove_empty_rows}
\title{Remove empty rows}
\usage{
remove_empty_rows(dt)
}
\arguments{
\item{df}{A data frame}
}
\value{
a packed df
}
\description{
\code{remove_empty_rows} finds rows with entire NAs and and wipes off.
}
