% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{run_mappings_sp}
\alias{run_mappings_sp}
\title{Run mappings}
\usage{
run_mappings_sp(dt = NULL)
}
\arguments{
\item{dt}{A list of data.tables or data.frames}
}
\value{
transformed list
}
\description{
\code{run_mappings_sp} run mapping scripts on the service provision sheets.
}
