% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{run_validation_sp}
\alias{run_validation_sp}
\title{Run validation}
\usage{
run_validation_sp(dt = NULL)
}
\arguments{
\item{dt}{A list of data.tables or data.frames}
}
\value{
A validated list
}
\description{
\code{run_validation_sp} runs validation scripts in the service provision sheets.
}
