% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{select_att}
\alias{select_att}
\title{Select Colums to Transform}
\usage{
select_att(dt = NULL, cols = NULL)
}
\arguments{
\item{dt}{a data.table or a list of data.frames}
}
\value{
a cleaned dataset
}
\description{
\code{select_att} selects columns names and types to transform from the attendance sheets.
}
