% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chop.R
\name{select_att_mma}
\alias{select_att_mma}
\title{Select Columns to Transform - MMA Attendance}
\usage{
select_att_mma(dt = NULL, cols = NULL)
}
\arguments{
\item{dt}{A data.table or data.frame}

\item{cols}{A vector of characters}
}
\value{
A chopped data.table or data.frame
}
\description{
\code{select_att_mma} selects columns names and types to transform from the MMA attendance sheets
}
\details{
It selects the following columns: State, Region, LGA, Ward, Facility, Girl ID, Age, Session 1 Date
and Attendance Frequency.
}
