% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chop.R
\name{select_prov}
\alias{select_prov}
\title{Select Columns to Transform}
\usage{
select_prov(dt = NULL, cols = NULL)
}
\arguments{
\item{dt}{A data.table or data.frame}

\item{cols}{A vector of characters}
}
\value{
a chopped data.table or data.frame
}
\description{
\code{select_prov} selects column names and types to transform from the service provison sheets.
}
\details{
It selects the following columns: State, Region, LGA , Ward, Facility, Type of Facility, Date Calculation,
ID #, Age (Formula), Program Entry Point, Current Method,Used EC/Condoms last sex, Pregnant?, Method recieved
and Received Condoms as a dual Method`
}
