% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_files.R
\name{to_r_date}
\alias{to_r_date}
\title{Covert excel dates to R dates}
\usage{
to_r_date(x, origin = "1900-01-01", ...)
}
\arguments{
\item{x}{Vector with xlsx dates numbers}

\item{origin}{Origin date 1900-01-01}
}
\value{
R date type object
}
\description{
\code{to_r_date} converts xlsx date number to R date
}
