% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transform_att}
\alias{transform_att}
\title{Transform elements}
\usage{
transform_att(dt)
}
\arguments{
\item{dt}{A list of data.table or a data.frames}
}
\value{
A list of transformed sheets.
}
\description{
Transform Attendance records
}
\details{
\code{transform_att} transforms attendance sheets to a standard format.
}
