% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transform_date}
\alias{transform_date}
\title{Transform eventDate}
\usage{
transform_date(dt)
}
\arguments{
\item{dt}{A data.table or list of data.frames}
}
\value{
A tranformed list
}
\description{
format dates to DHIS2 stardard event dates
}
