% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transform_sheets}
\alias{transform_sheets}
\title{Transform elements}
\usage{
transform_sheets(dt)
}
\arguments{
\item{dt}{A list of data.table or a data.frames}
}
\value{
A list of transformed sheets.
}
\description{
Transform sheets
}
\details{
\code{transform_sheets} transforms attendance sheets to a standard format.
}
