% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/push.R
\name{upload}
\alias{upload}
\title{Upload a payload}
\usage{
upload(
  df,
  live = FALSE,
  importStrategy = "CREATE",
  dryRun = F,
  skipNotifications = F
)
}
\arguments{
\item{df}{A data.frame object, the event payload.}

\item{importStrategy}{A string specifying the action to apply on the import.It provides options to Save objects of all,
new or update import status on the server. One of 'CREATE', 'UPDATE', 'CREATE_AND_UPDATE' and 'DELETE'.The default one is
"CREATE".}

\item{dryRun}{A logical input; Save changes on the server or just return the import summary.By default, The default one is
FALSE, to save the changes on the server.}

\item{skipNotifications}{A logical input; provides an option to send notifications for completed events. The default one is FALSE}
}
\value{
An S3 type object with content, path and the response.
}
\description{
This function uploads events to PSI data server, `data.psi-mis.org`.
}
\details{
\code{upload} function uploads individual events to two PSI servers; `data.psi-mis.org`
and `clone.psi-mis.org`. By default it sends the events to the test server available at `clone.psi-mis.org`.
}
\examples{
\dontrun{
# upload to test erver
upload(x)

# upload to production
upload(x, live = T)

# upload from a list
  upload_sheets <- function(x, live = F) {
x \%>\%
 map(., function(y) upload(y, live))
}

upload_sheets(x)

}
}
