% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_condom_as_dual}
\alias{validate_condom_as_dual}
\title{Validate condom as dual method}
\usage{
validate_condom_as_dual(x = NULL)
}
\value{
An existing option in DHIS2
}
\description{
\code{validate_condom_as_dual}  Validates the different options listed in the service provision sheets
against the options in NG RH A360 - Provider Client Records program in DHIS2.
}
