% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_current_method}
\alias{validate_current_method}
\title{Validate current method}
\usage{
validate_current_method(x = NULL)
}
\arguments{
\item{x}{A vector of character string}
}
\value{
An existing method in DHIS2
}
\description{
\code{validate_current_method} validates different current methods listed in the service provision sheets
against those provided in NG RH A360 - Provider Client Records program in DHIS2.
}
