% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_ec_last_sex}
\alias{validate_ec_last_sex}
\title{Validate used EC/Condom last sex}
\usage{
validate_ec_last_sex(x = NULL)
}
\arguments{
\item{x}{A vector of character string}
}
\value{
An existing option in DHIS2
}
\description{
\code{validate_ec_last_sex} validates the different options listed on service provision sheets against those provided
in NG RH A360 - Provider Client Records program in DHIS2.
}
