% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_method_recieved}
\alias{validate_method_recieved}
\title{Validate method recieved}
\usage{
validate_method_recieved(x = NULL)
}
\arguments{
\item{x}{A vector of character or string}
}
\description{
\code{validate_method_recieved} validates the different method options listed in the service provision sheets
against those provided in NG RH A360 - Provider Client Record program.
}
