% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_preg_results}
\alias{validate_preg_results}
\title{Validate pregancy results}
\usage{
validate_preg_results(x = NULL)
}
\arguments{
\item{x}{A vector of character string}
}
\value{
An existing option in DHIS2
}
\description{
\code{validate_preg_results} validates the pregnacy results in the service provision sheets against those provided
in the NG RH A360 - Provider Client Records program inn DHIS2.
}
