% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_program_activity}
\alias{validate_program_activity}
\title{Validate Program Activity}
\usage{
validate_program_activity(x = NULL)
}
\arguments{
\item{x}{A vector of character or string}
}
\value{
AN existing program activity in DHIS2
}
\description{
\code{validate_program_activity} validates the different program activities provided in the attendance sheets, including
MMA attendance against those listed in NG RG A360 - 9ja & MMA program.
}
