% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_program_entry_point}
\alias{validate_program_entry_point}
\title{Validate Program Entry Point}
\usage{
validate_program_entry_point(x = NULL)
}
\arguments{
\item{x}{A vector of character or string}
}
\value{
An existing program entry point in DHIS2
}
\description{
\code{validate_program_entry_point} validates the different program entry points provided in the service provision sheets
against those listed in NG RH A360 - Provider Client Record program.
}
