% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/option_set.R
\name{de_options}
\alias{de_options}
\title{Retrieve options in an option set}
\usage{
de_options(uid = NULL)
}
\arguments{
\item{uid}{A character string, uid of the option set.}
}
\value{
an S3 type object with content, path and the response.
}
\description{
Retrieve options in an option set
}
\details{
\code{de_options} is used internally by de_option_set to retreive options of an option set. It selects
    the name, code and id of options.
}
