% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/payload.R
\name{generate_payload}
\alias{generate_payload}
\title{Generate payload}
\usage{
generate_payload(dt, program_id = "program_id", status = "COMPLETED",
  storedBy = "old / archive import")
}
\arguments{
\item{dt}{A list of tranformed df}
}
\value{
list payload
}
\description{
Generate payload
}
