% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_files.R
\name{load_files}
\alias{load_files}
\title{Load A360 NG legacy files into R}
\usage{
load_files(path = NULL)
}
\arguments{
\item{path}{System path to the directory with data}
}
\value{
a list of data frames
}
\description{
\code{load_files} loads A360 legacy files in the specified directory into R
}
\examples{
\dontrun{
load_files(file.path(Sys.getenv("HOME"),"data")
}
}
