% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chop.R
\name{select_att}
\alias{select_att}
\title{Select Colums to Transform}
\usage{
select_att(dt = NULL, cols = NULL)
}
\arguments{
\item{dt}{A data.table or data,.frame}

\item{cols}{A vector of characters}
}
\value{
A chopped data.table or data.frame
}
\description{
\code{select_att} selects columns names and types to transform from the attendance sheets.
}
\details{
It selects the following columns: State, Region, LGA, Ward, Facility, Type of Facility, Name, ID #,
Age, Date Calculation and Program Activity.
}
