% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{select_prov}
\alias{select_prov}
\title{Select Columns to Transform}
\usage{
select_prov(dt = NULL, cols = NULL)
}
\description{
\code{select_prov} select column names and types to transform from the serv ice provison sheets.
}
